+++ llama.cpp/embedfile/embedfile.h
@@ -1,0 +1,6 @@
+#ifndef EMBEDFILE_H
+#define EMBEDFILE_H
+
+#define EMBEDFILE_VERSION "0.0.1-alpha.1"
+
+#endif
