+++ llama.cpp/embedfile/sqlite-lembed.h
