+++ llama.cpp/embedfile/sqlite-vec.c
