+++ llama.cpp/embedfile/sqlite-vec.h
