--- llama.cpp/embedr/embedr.h
@@ -1,6 +1,0 @@
-#ifndef EMBEDR_H
-#define EMBEDR_H
-
-#define EMBEDR_VERSION "0.0.1-alpha.1"
-
-#endif
